vol/requirements.txt
Django==1.11.20
Scrapy==1.5.2
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
scrapyd==1.2.0
jsonpath-rw==1.4.0
kombu==3.0.37
Celery[redis]==3.1.25
django-celery==3.2.1
future==0.17.1
pillow==5.2
attrs==18.2.0
flower==0.9.2
psycopg2==2.7.4
django-dynamic-scraper==0.13.2